set FALSE_ALARM boolean for testing purposes

+++ main.js
@@ -24,6 +24,7 @@ const SIMILARITY_THRESHOLD = config.text_analysis.max_similarity;
 const MAX_RETRIES_PER_FETCH = 3; //to be managed by user configuration
 const INITIAL_DEALY = 750; //to be managed by user configuration
 let FIFTEEN_MINUTES = 15 * 60000;
+let FALSE_ALARM = false;
 let BROWSER_PATH;
 
 const STRING_PLACEHOLDER = "placeholder";
@@ -730,7 +731,7 @@ const saveResults = async (results) => {
     const resultsWithTop = { results, topArticles, mostCommonTerm };
 
     fs.writeFileSync(resultsPath, JSON.stringify(resultsWithTop, null, 2));
-    if (thisIsTheTime) {
+    if (thisIsTheTime && !(FALSE_ALARM)) {
         fs.writeFileSync(flagPath, CRAWL_COMPLETE_TEXT);
         console.log("Crawling complete!")
     }
@@ -965,6 +966,7 @@ process.on('SIGINT', async () => {
     console.log(`Caught interrupt signal (Ctrl+C)\nSetting emailEndTime in ${(FIFTEEN_MINUTES/1000) - 1} minutes from now`);
     const now = new Date();
     emailEndTime = new Date(now.getTime() + FIFTEEN_MINUTES - 60000);
+    FALSE_ALARM = true;
 });
 
 const main = async () => {
@@ -985,7 +987,7 @@ const main = async () => {
         await saveResults(resultados);
     }
 
-    if (!fs.existsSync(path.join(__dirname, CRAWL_COMPLETE_FLAG))) {
+    if (!fs.existsSync(path.join(__dirname, CRAWL_COMPLETE_FLAG)) && !(FALSE_ALARM)) {
         fs.writeFileSync(path.join(__dirname, CRAWL_COMPLETE_FLAG), CRAWL_COMPLETE_TEXT);
     }
